Continuing on the UV import.
Fixed a bug that caused meshes with multiple UVs to fail to load.
The default argument technique for the current UV set name didn't work
in the original code, but the code always passed the UV name. I removed
the setter since nothing was using it, and added the set node name to
the setUVs call.

Cleaned up the code a little more by replacing boilerplate code with a
loop. Cleaned up some commented out cruft.