Added shader construction code.
Unlike Blender, Maya really wants to just get texture information
directly from files. We resolve this by generating the derived textures
in the import process, and then build the shaders to refer to the new
files we wrote. This is providing results that look consistent with what
comes from Blender. A good example for comparison is the Doomfist Formal
skin.