Convert default units to cm.
The usage of inches was to workaround the discrepancy that SEAnim files
had their units in inches due to the OW tools working around a quirk in
the Blender SEAnim importer.

For now, I'm living with a hack in SETools to account for this, and not
negatively impact the other users of the importer.