Fix textureList semantics to work with cached shaders.
When a shader was cached, the texture list wasn't properly associated
with it. This caused problems loading maps since that texture list is
scanned for matches against 'internal' textures. Objects with that
texture shouldn't be displayed, but this bug was causing false
negatives.

To avoid a global cache, the texture list is directly attached to the
shader node.