Fix material instancing.
I thought the ids baked into the models were globally unique. They are
not. So materials are now cached by Maya where we compute the material
name as a tuple of (modelID, materialID) which will be globally unique.
This detection is done in import_owmat and callers just have to provide
the model ID as the prefix argument.

This still results in Busan loading quite a bit quicker than before
since we're not loading any more data than necessary.